% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estab.R
\name{estab}
\alias{estab}
\title{Bioentity establishment or not at each node}
\usage{
estab(
  decvec,
  dispvec,
  estpm,
  estpsd,
  efmn,
  efsd,
  plotmp = F,
  xymat,
  readpestab,
  pestab
)
}
\arguments{
\item{decvec}{vector of 1=management, 0=no management}

\item{dispvec}{vector of 1=sp has dispersed to node, 0=sp has not dispersed to node}

\item{estpm}{mean probability of establishment (new or CONTINUED) in absence of management (used if \code{readpestab} =  F)}

\item{estpsd}{sd in probability of establishment in absence of management (used if \code{readpestab} = F)}

\item{efmn}{mean effect of management (proportional reduction in estabp, where the combination of efmn = 1 and efsd = 0 makes establishment impossible)}

\item{efsd}{sd of management effect}

\item{plotmp}{if T, then map of establishment is plotted}

\item{xymat}{matrix of x,y coordinates of nodes (used if \code{plotmp} = T)}

\item{readpestab}{if T, then a vector \code{pestab} is read in; otherwise the vector is generated using \code{estpm} and \code{estpsd}}

\item{pestab}{vector of probabilities of establishment (read in or generated when the function is run, depending on \code{readpestab} equal to T or F)}
}
\description{
This function determines whether a bioentity establishes or not at each node.  It is similar in some ways to the function \code{makedec}, which determines whether management is adopted.
}
\details{
Updated 2020-07-07
}
\examples{
x6 <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), xymat=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readpestab=F, estpm=0.5, estpsd=0.1, efmn=0.1, efsd=0.1, plotmp=T)
x7 <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), xymat=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readpestab=F, estpm=0.5, estpsd=0.1, efmn=0.9, efsd=0.1, plotmp=T)
x8 <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), xymat=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readpestab=T, pestab=c(0.9,0.1,0.1,0.1,0.1,0.9), efmn=0.9, efsd=0.1, plotmp=T)
x9 <- estab(decvec=c(1,1,1,0,0,0), dispvec=c(1,1,1,1,1,0), xymat=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T, ncol=2), readpestab=T, pestab=c(0.9,0.1,0.1,0.1,0.1,0.9), efmn=0.1, efsd=0.1, plotmp=T)
}
\keyword{establishment}
