% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlocs.R
\name{genlocs}
\alias{genlocs}
\title{Generate geographic locations of nodes in network}
\usage{
genlocs(extx, exty, nn, rand = TRUE)
}
\arguments{
\item{extx}{range of x coordinates}

\item{exty}{range of y coordinates}

\item{nn}{the number of nodes}

\item{rand}{if TRUE then locations are randomly generated (other options for distributing locations will be added)}
}
\description{
This function generates the geographic locations of nodes in network, for simulation studies where geographic locations are not observed.
}
\details{
Updated 2020-06-02
}
\examples{
xymat99 <- genlocs(extx=c(0,50), exty=c(0,50), nn=100, rand=TRUE)
}
\keyword{geography}
\keyword{locations}
