% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initvals.R
\name{initvals}
\alias{initvals}
\title{Generate initial status for nodes}
\usage{
initvals(xymat, loctype, ip, numinit, numorprop, plotmp = F)
}
\arguments{
\item{xymat}{n by 2 matrix of geographic coordinates, provided for a particular system or generated by function genlocs}

\item{loctype}{the type of locations where initial presence occurs: 'random' indicates all equally likely, 'upedge' indicates that nodes closest to the upper edge have presence, 'rightedge' indicates that nodes closest to the right edge have presence}

\item{ip}{the proportion of initial locations for presence}

\item{numinit}{the number of initial locations for presence}

\item{numorprop}{'num' indicates initial number for presence, 'prop' indicates initial proportion}

\item{plotmp}{if T, then map of presence and absence is plotted}
}
\description{
This function generates the initial status for nodes, in terms of the presence of management information or focal species
}
\details{
Updated 2020-06-02
}
\examples{
x2 <- initvals(j2 <- genlocs(extx=c(0,50), exty=c(0,50), nn=100, rand=TRUE),loctype='random', ip=0.1, numorprop='prop', plotmp=T)
x3 <- initvals(j2 <- genlocs(extx=c(0,50), exty=c(0,50), nn=100, rand=TRUE),loctype='upedge', ip=0.3, numorprop='prop', plotmp=T)
x4 <- initvals(j2 <- genlocs(extx=c(0,50), exty=c(0,50), nn=100, rand=TRUE), loctype='rightedge', ip=0.3, numorprop='prop',plotmp=T)
x5 <- initvals(matrix(c(1,5, 2,5, 3,5, 4,5), byrow=T, ncol=2), loctype='upedge', numorprop='num', numinit=2, plotmp=T) 
x6 <- initvals(matrix(c(5,1, 5,2, 5,3, 5,4), byrow=T, ncol=2), loctype='rightedge', numorprop='num', numinit=2, plotmp=T) 
}
\keyword{conditions}
\keyword{initial}
