% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedec.R
\name{makedec}
\alias{makedec}
\title{Decisions at each node about whether to adopt management}
\usage{
makedec(comvec, xymat, adpm, adpsd, readpadopt, padopt, plotmp = F)
}
\arguments{
\item{comvec}{vector of 1=info is present, 0=info is not present}

\item{xymat}{matrix of x,y coordinates for the nodes, for mapping if \code{plotmp} = T (and for determining the number of nodes in current version)}

\item{adpm}{mean probability of adopting management if informed}

\item{adpsd}{sd for drawing probability of adoption, in truncated normal distribution}

\item{readpadopt}{if T, then a VECTOR of padopt values is read in; if F, \code{adpm} and \code{adpsd} are used to generate the vector}

\item{padopt}{vector of probabilities of adoption if informed for nodes in the network (\code{readpadopt} determines whether \code{padopt} is read in to \code{makedec} or generated within \code{makedec} based on \code{adpm} and \code{adpsd})}

\item{plotmp}{if T, then map of decision is plotted}
}
\description{
This function generates decisions at each node about management
}
\details{
Updated 2020-06-03
}
\examples{
x5 <- makedec(comvec=c(1,1,1,0,0,0), xymat=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T,ncol=2), readpadopt=F, adpm=0.1, adpsd=0.1, plotmp=T)
x6 <- makedec(comvec=c(1,1,1,0,0,0), xymat=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T,ncol=2), readpadopt=F, adpm=0.5, adpsd=0.1, plotmp=T)
x7 <- makedec(comvec=c(1,1,1,0,0,0), xymat=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3), byrow=T,ncol=2), readpadopt=T, padopt=c(0.1,0.1,0.9,0.1,0.2,0.3), plotmp=T)
}
\keyword{decisions}
