% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisimtest.R
\name{multisimtest}
\alias{multisimtest}
\title{Evaluate nodes for value in invasion detection (based on a specified number of realizations)}
\usage{
multisimtest(adjmat, stoch, nrealz = 1)
}
\arguments{
\item{adjmat}{adjacency matrix for evaluation}

\item{stoch}{logical var indicating whether adjacency matrix entries are fixed or probabilities}

\item{nsim}{number of simulations to be analyzed}
}
\description{
Evaluate nodes for utility in invasion detection (based on a specified number of realizations) using the function multistart (which in turn uses the function onestart).  outarr saves the output from each realization from multistart, in a 3D array.  In meanarr, rows = starting nodes (introduction nodes), columns = sampling nodes, and entries = mean nodes free from invasion when invasion starting at the introduction node reaches the sampling node.  In vararr, for the same rows and columns, entries are the variance in the number of nodes free from invasion.  Note that there is no reason to use multisimtest if there is no stochastic component.
}
\examples{
multisimtest(adjmat=sAmat, stoch=T, nsim=10)[
}
\keyword{prioritization}
\keyword{sampling}
