% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistart.R
\name{multistart}
\alias{multistart}
\title{Evaluate nodes for value in invasion detection (invasion starting from each node in turn)}
\usage{
multistart(adjmat, wtvec, stoch = T)
}
\arguments{
\item{adjmat}{adjacency matrix for evaluation}

\item{wtvec}{vector of weights corresponding to nodes associated with adjacency matrix adjmat2}

\item{stoch}{logical var indicating whether adjacency matrix entries are fixed or probabilities}
}
\description{
For an analysis of each potential starting node in turn, generates a table of number of nodes infected by time detected at each potential sampling node, with rows being starting node and columns being sampling node - for one simulation.  Each node can be weighted with a relative likelihood of spread to that node.
}
\examples{
multistart(adjmat=Amat,wtvec= ___)
multistart(adjmat=sAmat,wtvec=____, stoch=T)
}
\keyword{prioritization}
\keyword{sampling}
