% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startwt.R
\name{startwt}
\alias{startwt}
\title{Evaluate nodes for value in invasion detection (based on a specified number of simulations, and including specified probabilities associated with each potential starting node)}
\usage{
startwt(msf.out, adjmat, wtvec, nodenam = NA)
}
\arguments{
\item{msf.out}{output object from function multisimf}

\item{adjmat}{adjacency matrix for evaluation}

\item{wtvec}{vector of weights indicating the probability that each node would be the starting node for invasion}

\item{nodenam}{vector of node names}
}
\description{
Uses weights that indicate how likely each node was to be the starting location for an invasion.  Uses output from multisimtest.  Finds the rate at which sampling failure (invasion of other nodes prior to detection at sampling node) is expected to occur for each potential sampling node.
}
\examples{
wtvec.ex <- c(1,10,100,1)
msf.outex <- multisimtest(adjmat=sAmat, stoch=T, nsim=10)
startwt(msf.out=msf.outex, adjmat=sAmat, wtvec=wtvec.ex)
startwt(msf.out=msf.outex, adjmat=sAmat, wtvec=wtvec.ex, nodenam=c("KS","NE","ND","SD"))
startwt(msf.out=msf.outex, adjmat=sAmat, wtvec=c(1,100,10,1), nodenam=c("KS","NE","ND","SD"))
}
\keyword{prioritization}
\keyword{sampling}
\keyword{weights}
