% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INAscene.R
\name{INAscene}
\alias{INAscene}
\title{Evaluates scenarios in an impact network analysis (INA)}
\usage{
INAscene(
  nreals,
  ntsteps,
  doplot = F,
  readgeocoords,
  geocoords = NA,
  numnodes = NA,
  xrange = NA,
  yrange = NA,
  randgeo = NA,
  readinitinfo,
  initinfo = NA,
  initinfo.norp = NA,
  initinfo.n = NA,
  initinfo.p = NA,
  initinfo.dist = NA,
  readinitbio,
  initbio = NA,
  initbio.norp = NA,
  initbio.n = NA,
  initbio.p = NA,
  initbio.dist = NA,
  readseam,
  seam = NA,
  seamdist = NA,
  seamrandp = NA,
  seampla = NA,
  seamplb = NA,
  seamlinktype = NA,
  seamthresh = NA,
  seamdiag1 = NA,
  seamformat = NA,
  seammax1 = NA,
  readbpam,
  bpam = NA,
  bpamdist = NA,
  bpamrandp = NA,
  bpampla = NA,
  bpamplb = NA,
  bpamlinktype = NA,
  bpamthresh = NA,
  bpamdiag1 = NA,
  bpamformat = NA,
  bpammax1 = NA,
  readprobadoptvec,
  probadoptvec = NA,
  probadoptmean = NA,
  probadoptsd = NA,
  readprobestabvec,
  probestabvec = NA,
  probestabmean = NA,
  probestabsd = NA,
  maneffmean = NA,
  maneffsd = NA,
  usethreshman,
  maneffthresh = NA,
  sampeffort = NA
)
}
\arguments{
\item{nreals}{number of realizations to be evaluated}

\item{ntsteps}{number of time steps to be evaluated}

\item{doplot}{if true, plots of resulting presence of information and bioentity are generated}

\item{readgeocoords}{if T, read in geocoords - otherwise, generate it in each realization}

\item{geocoords}{matrix of x,y coordinates of nodes}

\item{numnodes}{(genlocs:nodenum) the number of nodes, can be a vector of different numbers of nodes for scenario comparisons}

\item{xrange}{(genlocs:extx) range of x coordinates, e.g., c(0,50)}

\item{yrange}{(genlocs:exty) range of y coordinates, e.g., c(0,20)}

\item{randgeo}{(genlocs:rand) if TRUE then locations are randomly generated (the only location generation option for the moment)}

\item{readinitinfo}{info (setup2) if T, the initial values for the vector of starting locations for the presence of information are read in rather than generated}

\item{initinfo}{info (setup2) the vector of initial values read in if readinitinfo == T, a vector with lenth equal to the number of nodes and entries 1s or 0s with 1s indicating the initial presence of information}

\item{initinfo.norp}{info (initvals:numorpropi) 'num' indicates initial number for presence, 'prop' indicates initial proportion}

\item{initinfo.n}{info (initvals:numiniti) the number of initial locations for presence}

\item{initinfo.p}{info (initvals:propiniti) the proportion of initial locations for presence}

\item{initinfo.dist}{info (initvals:loctypei) the type of locations where initial presence occurs: 'random' indicates all equally likely, 'upedge' indicates that nodes closest to the upper edge have presence, 'rightedge' indicates that nodes closest to the right edge have presence}

\item{readinitbio}{bio (setup2) if T, the initial values for the vector of starting locations for the presence of the bioentity are read in rather than generated}

\item{initbio}{bio (setup2) the vector of initial values read in if readinitbio.s == T, a vector with lenth equal to the number of nodes and entries 1s or 0s with 1s indicating the initial presence of the bioentity}

\item{initbio.norp}{estab (initvals:numorprope/b) 'num' indicates initial number for presence, 'prop' indicates initial proportion}

\item{initbio.n}{estab (initvals:numinite/b) the number of initial locations for presence}

\item{initbio.p}{estab (initvals:propinite/b) the proportion of initial locations for presence}

\item{initbio.dist}{estab (initvals:loctypee/b) the type of locations where initial presence occurs: 'random' indicates all equally likely, 'upedge' indicates that nodes closest to the upper edge have presence, 'rightedge' indicates that nodes closest to the right edge have presence}

\item{readseam}{if T, the socioeconomic network adjacency matrix is read in rather than being generated from the outset}

\item{seam}{socioeconnomic network adjacency matrix, read in if readseam=T}

\item{seamdist}{com (genmovnet:distfcn) the function of distance used to estimate movement probability - 'random' (not related to distance) or 'powerlaw'}

\item{seamrandp}{(genmovnet) probability of link existence in random network generated for the socioeconomic network}

\item{seampla}{com (genmovnet:placn) power law parameter a in ad^(-b)}

\item{seamplb}{com (genmovnet:pllbcn) power law parameter b in ad^(-b)}

\item{seamlinktype}{com (genmovnet:lktypecn) link type, pa is presence/absence (unweighted, occurence/non-occurence), pr is a probability of occurence, wtd1 is general weight}

\item{seamthresh}{com (genmovnet:tlinkcn) threshold for whether link exists (probability in some cases)}

\item{seamdiag1}{com (spreadstep:diag1cn) if T, diagonal of adj mat is given value 1}

\item{seamformat}{com (spreadstep:mattypecn) 'fromto' indicates that rows are sources and columns are sinks, 'tofrom' indicates that columns are sources and rows are sinks}

\item{seammax1}{com (spreadstep:max1cn) if T, values in vect2 greater than 1 are replaced by 1}

\item{readbpam}{if T, the biophysical network adjacency matrix, describing dispersal likelihoods, is read in rather than being generated from the outset}

\item{bpam}{biophysical adjacency matrix, read in if readbpam=T}

\item{bpamdist}{disp (genmovnet:distfdn) the function of distance used to estimate movement probability - 'random' (not related to distance) or 'powerlaw'}

\item{bpamrandp}{(genmovnet) if bpamdist='random', the probability a link exists in the biophysical network adjacency matrix}

\item{bpampla}{disp (genmovnet:pladn) power law parameter a in ad^(-b)}

\item{bpamplb}{disp (genmovnet:plbdn) power law parameter b in ad^(-b)}

\item{bpamlinktype}{disp (genmovnet:lktypedn) link type, pa is presence/absence (unweighted, occurence/non-occurence), pr is a probability of occurence, wtd1 is general weight}

\item{bpamthresh}{disp (genmovnet:tlinkdn) threshold for whether link exists (probability in some cases)}

\item{bpamdiag1}{disp (spreadstep:diag1dn) if T, diagonal of adj mat is given value 1}

\item{bpamformat}{disp (spreadstep:mattypedn) 'fromto' indicates that rows are sources and columns are sinks, 'tofrom' indicates that columns are sources and rows are sinks}

\item{bpammax1}{disp (spreadstep:max1dn) if T, values in vect2 greater than 1 are replaced by 1}

\item{readprobadoptvec}{if T, read in a vector of probabilities of adoption for each node in socioeconomic network - if F, generate this vector}

\item{probadoptvec}{vector of probabilities of adoption for each node in socioeconomic network, read in if readprobadoptvec=T}

\item{probadoptmean}{(makedec:adpmn) mean probability of adopting management if informed}

\item{probadoptsd}{(makedec:adpsdn) sd in truncated normal distribution of probability of adoption}

\item{readprobestabvec}{if T, read in a vector of probabilities of establishment for each node in biophysical network - if F, generate this vector}

\item{probestabvec}{vector of probabilities of establishment for each node in biophysical network, read in if readprobestabvec=T}

\item{probestabmean}{(estab:estpmn) mean probability of establishment (new or CONTINUED) in absence of management}

\item{probestabsd}{(estab:estpsdn) sd of probability of establishment in absence of management in truncated normal distribution}

\item{maneffmean}{(estinfo:efmn) the underlying mean change in establishment probability (as a proportion) resulting from management technology}

\item{maneffsd}{(estinfo) the standard deviation of the management technology effect}

\item{usethreshman}{if T, the threshold for management is used, so that information is never present anywhere unless the management effect estimate exceeds the threshold}

\item{maneffthresh}{(estinfo:efthr) the threshold effect size for communicating about management}

\item{sampeffort}{(estinfo) the sampling effort, where increasing sampling effort results in a more precise estimate of the management effect size}
}
\description{
This function implements and summarizes multiple simulations across a designated range of parameter values
}
\details{
Updated 2020-07-07
}
\examples{
library(truncnorm)
j25.readgeocoords <- INAscene(nreals=3, ntsteps=3, doplot=F, readgeocoords=T, geocoords=matrix(c(1,1, 1,2, 1,3, 2,1, 2,2, 2,3),byrow=T,ncol=2), numnodes=NA, xrange=NA, yrange=NA, randgeo=F, readinitinfo=T, initinfo=c(1,1,1,0,0,0), initinfo.norp=NA, initinfo.n=NA, initinfo.p=NA, initinfo.dist=NA, readinitbio=T, initbio=c(0,0,0,1,1,1), initbio.norp=NA, initbio.n=NA, initbio.p=NA,  initbio.dist=NA, readseam=F, seam=NA, seamdist='random', seamrandp=c(0.01,0.05,0.1,0.5), seampla=NA, seamplb=NA, seamlinktype='pa', seamthresh=NA, seamdiag1=T, seamformat='fromto', seammax1=T, readbpam=F, bpam=NA, bpamdist='random', bpamrandp=0.1, bpampla=NA, bpamplb=NA, bpamlinktype='pa', bpamthresh=NA, bpamdiag1=T, bpamformat='fromto', bpammax1=T, readprobadoptvec=F, probadoptvec=NA, probadoptmean=0.1, probadoptsd=0.1, readprobestabvec=F, probestabvec=NA, probestabmean=0.1, probestabsd=0.1, maneffmean=0.5, maneffsd=0.1, usethreshman=F, maneffthresh=NA, sampeffort=NA) 
j25.baseline <- INAscene(nreals=3, ntsteps=3, doplot=F, readgeocoords=F, geocoords=NA, numnodes=50, xrange=c(0,10), yrange=c(0,20), randgeo=T, readinitinfo=F, initinfo=NA, initinfo.norp='num', initinfo.n=5, initinfo.p=NA, initinfo.dist='random', readinitbio=F, initbio=NA, initbio.norp='num', initbio.n=5, initbio.p=NA,  initbio.dist='random', readseam=F, seam=NA, seamdist='powerlaw', seamrandp=NA, seampla=1, seamplb=0.5, seamlinktype='pa', seamthresh=0.1, seamdiag1=T, seamformat='fromto', seammax1=T, readbpam=F, bpam=NA, bpamdist='powerlaw', bpamrandp=NA, bpampla=1, bpamplb=0.5, bpamlinktype='pa', bpamthresh=0.5, bpamdiag1=T, bpamformat='fromto', bpammax1=T, readprobadoptvec=F, probadoptvec=NA, probadoptmean=0.5, probadoptsd=0.2, readprobestabvec=F, probestabvec=NA, probestabmean=0.5, probestabsd=0.2, maneffmean=0.5, maneffsd=0.2, usethreshman=T, maneffthresh=0.5, sampeffort=2) 
j25.baseline$multout
sens.probadoptmean <- INAscene(nreals=15, ntsteps=3, doplot=F, readgeocoords=F, geocoords=NA, numnodes=50, xrange=c(0,10), yrange=c(0,20), randgeo=T, readinitinfo=F, initinfo=NA, initinfo.norp='num', initinfo.n=5, initinfo.p=NA, initinfo.dist='random', readinitbio=F, initbio=NA, initbio.norp='num', initbio.n=5, initbio.p=NA,  initbio.dist='random', readseam=F, seam=NA, seamdist='powerlaw', seamrandp=NA, seampla=1, seamplb=0.5, seamlinktype='pa', seamthresh=0.1, seamdiag1=T, seamformat='fromto', seammax1=T, readbpam=F, bpam=NA, bpamdist='powerlaw', bpamrandp=NA, bpampla=1, bpamplb=0.5, bpamlinktype='pa', bpamthresh=0.5, bpamdiag1=T, bpamformat='fromto', bpammax1=T, readprobadoptvec=F, probadoptvec=NA, probadoptmean=seq(0,1,0.1), probadoptsd=0.2, readprobestabvec=F, probestabvec=NA, probestabmean=0.5, probestabsd=0.2, maneffmean=0.9, maneffsd=0.2, usethreshman=T, maneffthresh=0.3, sampeffort=2) 
jt <- sens.probadoptmean$multout
plot(jt$probadoptmean, jt$mestab, xlab='Mean probability of adopting technology if informed', ylab='Proportion nodes with bioentity')
plot(jt$probadoptmean, jt$mdec, xlab='Mean probability of adopting technology if informed', ylab='Proportion nodes with technology adoption')
sens.seamplb <- INAscene(nreals=15, ntsteps=3, doplot=F, readgeocoords=F, geocoords=NA, numnodes=50, xrange=c(0,50), yrange=c(0,50), randgeo=T, readinitinfo=F, initinfo=NA, initinfo.norp='num', initinfo.n=5, initinfo.p=NA, initinfo.dist='random', readinitbio=F, initbio=NA, initbio.norp='num', initbio.n=5, initbio.p=NA,  initbio.dist='random', readseam=F, seam=NA, seamdist='powerlaw', seamrandp=NA, seampla=1, seamplb=seq(0,2,0.1), seamlinktype='pa', seamthresh=0.1, seamdiag1=T, seamformat='fromto', seammax1=T, readbpam=F, bpam=NA, bpamdist='powerlaw', bpamrandp=NA, bpampla=1, bpamplb=0.5, bpamlinktype='pa', bpamthresh=0.5, bpamdiag1=T, bpamformat='fromto', bpammax1=T, readprobadoptvec=F, probadoptvec=NA, probadoptmean=0.7, probadoptsd=0.2, readprobestabvec=F, probestabvec=NA, probestabmean=0.5, probestabsd=0.2, maneffmean=0.9, maneffsd=0.2, usethreshman=T, maneffthresh=0.3, sampeffort=2) 
jt2 <- sens.seamplb$multout
plot(jt2$seamplb, jt2$mestab, xlab='Power law parameter b in ad^(-b) for communication links', ylab='Proportion nodes with bioentity')
plot(jt2$seamplb, jt2$mdec, xlab='Power law parameter b in ad^(-b) for communication links', ylab='Proportion nodes with technology adoption')
plot(jt2$seamplb, jt2$mcom, xlab='Power law parameter b in ad^(-b) for communication links', ylab='Proportion nodes with information about technology')
j25.seamrandp <- INAscene(nreals=3, ntsteps=3, doplot=F, readgeocoords=F, geocoords=NA, numnodes=50, xrange=c(0,10), yrange=c(0,20), randgeo=T, readinitinfo=F, initinfo=NA, initinfo.norp='num', initinfo.n=5, initinfo.p=NA, initinfo.dist='random', readinitbio=F, initbio=NA, initbio.norp='num', initbio.n=5, initbio.p=NA,  initbio.dist='random', readseam=F, seam=NA, seamdist='random', seamrandp=c(0.01,0.05,0.1,0.5), seampla=NA, seamplb=NA, seamlinktype='pa', seamthresh=NA, seamdiag1=T, seamformat='fromto', seammax1=T, readbpam=F, bpam=NA, bpamdist='random', bpamrandp=0.1, bpampla=NA, bpamplb=NA, bpamlinktype='pa', bpamthresh=NA, bpamdiag1=T, bpamformat='fromto', bpammax1=T, readprobadoptvec=F, probadoptvec=NA, probadoptmean=0.1, probadoptsd=0.1, readprobestabvec=F, probestabvec=NA, probestabmean=0.1, probestabsd=0.1, maneffmean=0.5, maneffsd=0.1, usethreshman=F, maneffthresh=NA, sampeffort=NA) 
}
\keyword{experiments}
\keyword{simulation}
