% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisimtest.R
\name{multisimtest}
\alias{multisimtest}
\title{Evaluate nodes for value in invasion detection (based on a specified number of simulations)}
\usage{
multisimtest(adjmat, wtvec, stoch = T, nsim = 1)
}
\arguments{
\item{adjmat}{adjacency matrix for evaluation}

\item{stoch}{logical var indicating whether adjacency matrix entries are fixed or probabilities}

\item{nsim}{number of simulations to be analyzed}
}
\description{
For an analysis of each potential starting node in turn, outputs a table of number of nodes infected by time detected at each potential sampling node, with rows being starting node and columns being sampling node - for a specified number of simulation
}
\examples{
multisimtest(adjmat=Amat, wtvec=__, nsim=1)  # multiple sims not needed because not stochastic - BUT, yes, now should be stochastic
multisimtest(adjmat=sAmat, wtvec=__, stoch=T, nsim=10)
}
\keyword{prioritization}
\keyword{sampling}
